UTP-020-2: Rearrange files and remove unnecessary codes
